\name{tmR-package}
\alias{tmR-package}
\alias{tmR}
\docType{package}
\title{
Text mining in R
}
\description{
Make a more stable text mining function in every operation system.
}
\details{
\tabular{ll}{
Package: \tab tmR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-29\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Andrew Tang
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ text mining }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ 
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
x <- 1
}
